% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{Accuracy}
\alias{Accuracy}
\title{Estatisticas de acuracia}
\usage{
Accuracy(fit, dig = 4)
}
\arguments{
\item{fit}{objeto da classe forecast}

\item{dig}{número de digitos da saida}
}
\description{
Calcula as estatísticas "ME", "RMSE", "MAE", "MPE", "MAPE", "SMAPE", "LSD" e "LAR"
}

