% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{arimaForecast}
\alias{arimaForecast}
\title{Arima forecast wrapper}
\usage{
arimaForecast(x, h, level = 95, onlyfc = TRUE, xreg = NULL,
  newxreg = NULL, ...)
}
\description{
Arima forecast wrapper
}

