% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{lmForecast}
\alias{lmForecast}
\title{Linear model forecast wrapper}
\usage{
lmForecast(x, h, level = 95, onlyfc = TRUE, xreg = NULL, newxreg = NULL,
  ...)
}
\description{
Linear model forecast wrapper
}

