% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{naiveForecast}
\alias{naiveForecast}
\title{Naive forecast wrapper}
\usage{
naiveForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
Naive forecast wrapper
}

