% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{rwForecast}
\alias{rwForecast}
\title{Random walk forecast wrapper}
\usage{
rwForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
Random walk forecast wrapper
}

