% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{snaiveForecast}
\alias{snaiveForecast}
\title{Seasonal naive forecast wrapper}
\usage{
snaiveForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
Seasonal naive forecast wrapper
}

