% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{stl.Forecast}
\alias{stl.Forecast}
\title{Stl forecast wrapper}
\usage{
stl.Forecast(x, h, level = 95, method = "ets", onlyfc = TRUE, ...)
}
\description{
Stl forecast wrapper
}

