% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{tbatsForecast}
\alias{tbatsForecast}
\title{TBATS forecast wrapper}
\usage{
tbatsForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
TBATS forecast wrapper
}

