% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMISMultiForecastExtraFunctions.R
\name{thetaForecast}
\alias{thetaForecast}
\title{Theta forecast wrapper}
\usage{
thetaForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
Theta forecast wrapper
}

